# docs/swagger.yaml
definitions:
  dto.AddMediaGroupRequest:
    properties:
      group_id:
        format: uuid
        type: string
      relation_type:
        enum:
        - content
        - gallery
        - attachment
        type: string
    required:
    - group_id
    - relation_type
    type: object
  dto.AttachMediaRequest:
    properties:
      group_id:
        type: string
      media_id:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - group_id
    - media_id
    type: object
  dto.BlogPostListResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      posts:
        items:
          $ref: '#/definitions/dto.BlogPostResponse'
        type: array
      total_count:
        type: integer
    type: object
  dto.BlogPostResponse:
    properties:
      author_id:
        format: uuid
        type: string
      content:
        type: string
      created_at:
        type: string
      excerpt:
        type: string
      featured_image_id:
        format: uuid
        type: string
      featured_image_path:
        type: string
      id:
        format: uuid
        type: string
      media_groups:
        additionalProperties:
          items:
            $ref: '#/definitions/dto.MediaItemResponse'
          type: array
        type: object
      metadata:
        additionalProperties: {}
        type: object
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateBlogPostRequest:
    properties:
      author_id:
        format: uuid
        type: string
      content:
        type: string
      excerpt:
        maxLength: 255
        type: string
      featured_image_id:
        format: uuid
        type: string
      metadata:
        additionalProperties: {}
        type: object
      published_at:
        type: string
      slug:
        type: string
      status:
        enum:
        - draft
        - published
        - archived
        type: string
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - author_id
    - content
    - title
    type: object
  dto.CreateGalleryRequest:
    properties:
      author_id:
        type: string
      cover_image_index:
        description: Индекс изображения, используемого как обложка
        type: integer
      description:
        type: string
      images:
        description: Список изображений в галерее
        items:
          type: string
        type: array
      metadata:
        additionalProperties: true
        type: object
      slug:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - author_id
    - title
    type: object
  dto.MediaGroupResponse:
    properties:
      added_at:
        type: string
      group_id:
        format: uuid
        type: string
      relation_type:
        type: string
    type: object
  dto.MediaItemResponse:
    properties:
      group_id:
        type: string
      id:
        type: string
      position:
        type: integer
      storage_path:
        type: string
    type: object
  dto.PostMediaGroupsResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/dto.MediaGroupResponse'
        type: array
      post_id:
        format: uuid
        type: string
    type: object
  dto.UpdateBlogPostRequest:
    properties:
      content:
        type: string
      excerpt:
        maxLength: 255
        type: string
      featured_image_id:
        format: uuid
        type: string
      metadata:
        additionalProperties: {}
        type: object
      published_at:
        type: string
      slug:
        type: string
      status:
        enum:
        - draft
        - published
        - archived
        type: string
      title:
        maxLength: 100
        minLength: 3
        type: string
    type: object
  dto.UserRegisterInput:
    properties:
      email:
        type: string
      is_admin:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      phone:
        description: Формат +71234567890
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
  models.Media:
    properties:
      created_at:
        type: string
      duration:
        type: integer
      file_size:
        type: integer
      height:
        type: integer
      id:
        type: string
      is_public:
        type: boolean
      media_type:
        $ref: '#/definitions/models.MediaType'
      metadata:
        $ref: '#/definitions/models.Metadata'
      mime_type:
        type: string
      original_filename:
        type: string
      storage_path:
        type: string
      uploader_id:
        type: string
      width:
        type: integer
    type: object
  models.MediaType:
    enum:
    - photo
    - video
    - audio
    - document
    type: string
    x-enum-varnames:
    - MediaTypePhoto
    - MediaTypeVideo
    - MediaTypeAudio
    - MediaTypeDocument
  models.Metadata:
    additionalProperties: true
    type: object
  models.User:
    properties:
      basket_id:
        type: string
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      last_login:
        type: string
      name:
        type: string
      password:
        items:
          type: integer
        type: array
      phone:
        type: string
      registration_date:
        type: string
    type: object
  request.LoginRequest:
    properties:
      identifier:
        description: |-
          Email    string `json:"email,omitempty"`
          Phone    string `json:"phone,omitempty"`
        type: string
      password:
        minLength: 8
        type: string
    required:
    - identifier
    - password
    type: object
  response.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
      status:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Вход в систему по email и паролю. Возвращает JWT-токен.
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход (токен)
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Ошибка аутентификации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - users
  /api/v1/media/groups:
    post:
      consumes:
      - multipart/form-data
      description: Создает новую группу для организации медиафайлов
      parameters:
      - description: UUID владельца группы
        format: uuid
        in: formData
        name: owner_id
        required: true
        type: string
      - description: Описание группы
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Группа создана (no content)
        "400":
          description: Невалидный UUID владельца
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка создания группы
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать медиагруппу
      tags:
      - Медиа-группы
  /api/v1/media/groups/{group_id}:
    get:
      description: Возвращает список всех медиафайлов в группе
      parameters:
      - description: UUID группы
        format: uuid
        in: path
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список медиафайлов
          schema:
            items:
              $ref: '#/definitions/models.Media'
            type: array
        "400":
          description: Невалидный UUID группы
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка получения списка
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить медиа группы
      tags:
      - Медиа-группы
  /api/v1/media/groups/{group_id}/attach:
    post:
      consumes:
      - application/json
      description: Связывает один или несколько медиафайлов с существующей медиагруппой
      parameters:
      - description: UUID группы
        format: uuid
        in: path
        name: group_id
        required: true
        type: string
      - description: Данные для прикрепления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AttachMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное прикрепление (no content)
        "400":
          description: 'Невалидные данные: пустой массив, неверный формат UUID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Группа не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "413":
          description: Превышен лимит количества медиафайлов
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Ошибка привязки медиа
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Прикрепить медиа к группе
      tags:
      - Медиа-группы
  /api/v1/media/images:
    get:
      consumes:
      - application/json
      description: Возвращает список всех загруженных изображений с метаданными
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить все изображения
      tags:
      - Медиа
  /api/v1/media/upload:
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл на сервер с возможностью указания метаданных
      parameters:
      - description: Файл для загрузки (макс. 10MB)
        in: formData
        name: file
        required: true
        type: file
      - description: UUID пользователя-загрузчика
        format: uuid
        in: formData
        name: uploader_id
        required: true
        type: string
      - description: Тип контента
        enum:
        - photo
        - video
        - audio
        - document
        in: formData
        name: media_type
        required: true
        type: string
      - description: Публичный доступ (по умолчанию false)
        in: formData
        name: is_public
        type: boolean
      - description: Дополнительные метаданные в JSON-формате
        in: formData
        name: metadata
        type: string
      - description: Ширина в пикселях (для изображений/видео)
        in: formData
        name: width
        type: integer
      - description: Высота в пикселях (для изображений/видео)
        in: formData
        name: height
        type: integer
      - description: Длительность в секундах (для видео/аудио)
        in: formData
        name: duration
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Успешно загруженный медиаобъект
          schema:
            $ref: '#/definitions/models.Media'
        "400":
          description: Некорректные входные данные
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "413":
          description: Превышен максимальный размер файла
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "415":
          description: Неподдерживаемый тип файла
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузка медиафайла
      tags:
      - Медиа
  /api/v1/posts:
    get:
      description: Возвращает список постов с пагинацией и фильтрацией по статусу.
        http://localhost:8080/api/v1/posts?status=archived&page=1&per_page=1
      parameters:
      - description: Фильтр по статусу (draft, published, archived)
        in: query
        name: status
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogPostListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список постов
      tags:
      - Посты
    post:
      consumes:
      - application/json
      description: Создает новый пост блога. Добавлять только authod_id -> существующий
        пользователь. Добавлять FeaturedImageID -> только существующую медиа
      parameters:
      - description: Данные поста
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBlogPostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BlogPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать новый пост
      tags:
      - Посты
  /api/v1/posts/{id}:
    delete:
      description: Удаляет пост (физическое удаление)
      parameters:
      - description: UUID поста
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить пост
      tags:
      - Посты
    get:
      description: Возвращает пост по его ID
      parameters:
      - description: UUID поста
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogPostResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить пост
      tags:
      - Посты
    put:
      consumes:
      - application/json
      description: Обновляет данные поста
      parameters:
      - description: UUID поста
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBlogPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить пост
      tags:
      - Посты
  /api/v1/posts/{id}/archive:
    patch:
      description: Архивирует пост (soft delete)
      parameters:
      - description: UUID поста
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Архивировать пост
      tags:
      - Посты
  /api/v1/posts/{id}/media-groups:
    get:
      description: Возвращает список медиа-групп, привязанных к посту
      parameters:
      - description: UUID поста
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Тип связи (content, gallery, attachment)
        in: query
        name: relation_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostMediaGroupsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить медиа-группы поста
      tags:
      - Посты
    post:
      consumes:
      - application/json
      description: Привязывает медиа-группу к посту с указанием типа связи
      parameters:
      - description: UUID поста
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Данные медиа-группы
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddMediaGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostMediaGroupsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Добавить медиа-группу к посту
      tags:
      - Посты
  /api/v1/posts/{id}/publish:
    patch:
      description: Устанавливает статус поста "published"
      parameters:
      - description: UUID поста
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Опубликовать пост
      tags:
      - Посты
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Создание аккаунта. Возвращает ID пользователя.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    user_id:
                      type: string
                  type: object
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - users
  /api/v1/users/{user_id}/is-admin:
    get:
      consumes:
      - application/json
      description: Проверяет, является ли указанный пользователь администратором
      parameters:
      - description: UUID пользователя
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Результат проверки" example({"is_admin": true})'
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: 'Невалидный UUID" example({"error": "invalid user ID format"})'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Ошибка сервера" example({"error": "failed to check admin status"})'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Проверка административного статуса пользователя
      tags:
      - Users
  /api/v1/users/users_id:
    post:
      consumes:
      - application/json
      description: Возвращает полную информацию о пользователе по его UUID
      parameters:
      - description: UUID пользователя
        example: '"a8a8a8a8-a8a8-a8a8-a8a8-a8a8a8a8a8a8"'
        format: uuid
        in: body
        name: user_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешно полученные данные пользователя
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректный UUID пользователя
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение информации о пользователе
      tags:
      - Пользователи
  /galleries:
    post:
      consumes:
      - application/json
      description: Создает новую галерею на основе переданных данных.
      parameters:
      - description: Данные для создания галереи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGalleryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Галерея успешно создана, возвращает ID созданной галереи
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректные данные запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при создании галереи
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создание новой галереи
      tags:
      - Галереи
swagger: "2.0"